requirements.txt
djangorestframework
django-smart-selects
Django==3.1.3

